requirements.txt
psutil==5.9.5
customtkinter==5.2.0
pillow==10.0.0 